requirements.txt
deepspeed==0.13.1
nlp==0.4.0
nltk==3.8.1
openai==1.50.0
pandas==2.2.3
python-dotenv==1.0.1
rich==13.8.1
scikit_learn==1.5.2
scipy==1.14.1
shortuuid==1.0.13
tiktoken==0.7.0
torch==2.4.0
tqdm==4.66.4
transformers==4.44.2
vllm==0.5.5
